First commit for adding BT Audio
This is the first commit for adding Bluetooth audio to the kv4p radio.
The basic concept is to add a Bluetooth Serial Port Profile (SPP/RFCOMM)
for the Android <-> radio data link. The initial test was using the
BTSerial from the ESP32 and sending characters to the open source Serial
Bluetooth Terminal
(https://github.com/kai-morich/SimpleBluetoothTerminal,
https://play.google.com/store/apps/details?id=de.kai_morich.serial_bluetooth_terminal&hl=en_US) on Android.
Once this worked, I attempted to pull in the serial port handling
concepts from the Serial Bluetooth Terminal.